pyproject.toml
[tool.poetry]
name = "workshop kafka"
version = "0.1.0"
description = ""
authors = ["SantiagoLinux <santiago.gomez.castro.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
jupyter = "^1.1.1"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"
psycopg2 = "^2.9.10"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
